add ast_convert_test
