+++ chore/gogensig/convert/ast_convert.go
@@ -1,8 +1,6 @@
 package convert
 
 import (
-	"fmt"
-
 	"github.com/goplus/llgo/chore/gogensig/config"
 	"github.com/goplus/llgo/chore/gogensig/visitor"
 	"github.com/goplus/llgo/chore/llcppg/ast"
@@ -50,30 +48,27 @@ func (p *AstConvert) VisitFuncDecl(funcDecl *ast.FuncDecl) {
 	p.pkg.NewFuncDecl(funcDecl)
 }
 
+/*
+//TODO
 func (p *AstConvert) VisitClass(className *ast.Ident, fields *ast.FieldList, typeDecl *ast.TypeDecl) {
 	fmt.Printf("visit class %s\n", className.Name)
 	p.pkg.NewTypeDecl(typeDecl)
-	//TODO new struct and convert fields
 }
 
 func (p *AstConvert) VisitMethod(className *ast.Ident, method *ast.FuncDecl, typeDecl *ast.TypeDecl) {
 	fmt.Printf("visit method %s of %s\n", method.Name.Name, className.Name)
-	//TODO convert method decl
-}
+}*/
 
 func (p *AstConvert) VisitStruct(structName *ast.Ident, fields *ast.FieldList, typeDecl *ast.TypeDecl) {
 	p.pkg.NewTypeDecl(typeDecl)
 }
 
-func (p *AstConvert) VisitEnum(enumName *ast.Ident, fields *ast.FieldList, typeDecl *ast.TypeDecl) {
-	fmt.Printf("visit enum %s\n", enumName.Name)
-	//TODO convert enum
-}
-
+/*
+//TODO
 func (p *AstConvert) VisitUnion(unionName *ast.Ident, fields *ast.FieldList, typeDecl *ast.TypeDecl) {
 	//fmt.Printf("visit union %s\n", unionName.Name)
 	//TODO convert union
-}
+}*/
 
 func (p *AstConvert) VisitEnumTypeDecl(enumTypeDecl *ast.EnumTypeDecl) {
 	p.pkg.NewEnumTypeDecl(enumTypeDecl)
@@ -86,7 +81,5 @@ func (p *AstConvert) VisitTypedefDecl(typedefDecl *ast.TypedefDecl) {
 func (p *AstConvert) VisitDone(docPath string) {
 	if p.visitDone != nil {
 		p.visitDone(p.pkg, docPath)
-	} else {
-		p.pkg.Write(docPath, "")
 	}
 }
